+++ network/api_test.go
@@ -153,15 +153,21 @@ func TestMain(m *testing.M) {
 }
 
 func TestActivate(t *testing.T) {
+	const expectedManifest = string(`{"Implements":["` + ImplementationName + `"]}`)
+
 	response, err := http.Get("http://localhost:32234/Plugin.Activate")
 	if err != nil {
 		t.Fatal(err)
 	}
 	defer response.Body.Close()
 	body, err := io.ReadAll(response.Body)
+	if err != nil {
+		t.Fatal(err)
+	}
+	bodyStr := strings.TrimSpace(string(body))
 
-	if string(body) != manifest+"\n" {
-		t.Fatalf("Expected %s, got %s\n", manifest+"\n", string(body))
+	if bodyStr != expectedManifest {
+		t.Fatalf("Expected '%s', got '%s'\n", expectedManifest, bodyStr)
 	}
 }